requirement.txt
# ML & Image Processing
torch
torchvision
numpy
matplotlib
Pillow
tqdm
opencv-python

# Backend API
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
python-multipart
